softgpu: Move triangle rasterization code to a new file.
