+++ GPU/GPU.vcxproj
@@ -162,6 +162,7 @@
     <ClInclude Include="GPUState.h" />
     <ClInclude Include="Math3D.h" />
     <ClInclude Include="Null\NullGpu.h" />
+    <ClInclude Include="Software\Rasterizer.h" />
     <ClInclude Include="Software\SoftGpu.h" />
     <ClInclude Include="Software\TransformUnit.h" />
   </ItemGroup>
@@ -185,6 +186,7 @@
     <ClCompile Include="GPUState.cpp" />
     <ClCompile Include="Math3D.cpp" />
     <ClCompile Include="Null\NullGpu.cpp" />
+    <ClCompile Include="Software\Rasterizer.cpp" />
     <ClCompile Include="Software\SoftGpu.cpp" />
     <ClCompile Include="Software\TransformUnit.cpp" />
   </ItemGroup>
